not really working.

+++ CK3toEU4/Source/CK3World/Characters/CharacterDomain.h
@@ -20,8 +20,8 @@ class CharacterDomain: commonItems::parser
 	[[nodiscard]] const auto& getRealmCapital() const { return realmCapital; }
 	[[nodiscard]] const auto& getDomain() const { return domain; }
 
-	void loadRealmCapital(const std::pair<int, std::shared_ptr<Title>>& theRealmCapital) { realmCapital = theRealmCapital; }
-	void loadDomain(const std::vector<std::pair<int, std::shared_ptr<Title>>>& theDomain) { domain = theDomain; }
+	void loadRealmCapital(const std::pair<long long, std::shared_ptr<Title>>& theRealmCapital) { realmCapital = theRealmCapital; }
+	void loadDomain(const std::vector<std::pair<long long, std::shared_ptr<Title>>>& theDomain) { domain = theDomain; }
 
   private:
 	void registerKeys();
@@ -30,8 +30,8 @@ class CharacterDomain: commonItems::parser
 	double vassalStrength = 0;
 	std::string government;
 	std::set<std::string> laws;
-	std::pair<int, std::shared_ptr<Title>> realmCapital;			// A barony!
-	std::vector<std::pair<int, std::shared_ptr<Title>>> domain; // These are all titles owned (b-c-d-k-e), landless included. First one is PRIMARY.
+	std::pair<long long, std::shared_ptr<Title>> realmCapital;			// A barony!
+	std::vector<std::pair<long long, std::shared_ptr<Title>>> domain; // These are all titles owned (b-c-d-k-e), landless included. First one is PRIMARY.
 };
 } // namespace CK3
