+++ CK3toEU4/Source/CK3World/Religions/Faiths.h
@@ -19,7 +19,7 @@ class Faiths: commonItems::parser
   private:
 	void registerKeys();
 
-	std::map<int, std::shared_ptr<Faith>> faiths;
+	std::map<long long, std::shared_ptr<Faith>> faiths;
 };
 } // namespace CK3
